requirements.txt
discord
discord-components
pytz
mysql-connector
PyYAML
colorama